--- mcweb/frontend/src/Homepage.js
@@ -1,12 +1,0 @@
-
-import React from 'react'
-import Header from './Header';
-const Homepage = () => {
-  return (
-    <div className="homepage">
-      <Header />
-    </div>
-  );
-}
-
-export default Homepage
